+++ src/main/java/com/app/banking/data/sql/entity/CardType.java
@@ -32,5 +32,5 @@ public class CardType {
     private Integer maxWithdrawal;
 
     @Column(name = "cashbackPercent", nullable = false)
-    private Double cashbackPercent;
+    private Float cashbackPercent;
 }
