+++ src/main/java/com/app/banking/data/sql/repo/BusinessRepository.java
@@ -3,5 +3,5 @@ package com.app.banking.data.sql.repo;
 import com.app.banking.data.sql.entity.Business;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessRepository extends JpaRepository<Business, Long> {
+public interface BusinessRepository extends JpaRepository<Business, Integer> {
 }
